update best model file name

+++ main_from_HGP-SL.py
@@ -114,8 +114,9 @@ def train():
 
     files = glob.glob('*.pth')
     for f in files:
+        model_name = f.split('.')[0].split('_')[0]
         epoch_nb = int(f.split('.')[0].split('_')[-1])
-        if epoch_nb > best_epoch:
+        if epoch_nb > best_epoch and model_name == model_name_prefix:
             os.remove(f)
     print('Optimization Finished! Total time elapsed: {:.6f}'.format(time.time() - t))
